`_InternalTestSupport` should not depend on `Build`
Looks like something I caused a while ago in d457fa46b3 which was later further entrenched in `PluginInvocationTests`, but we should keep the "data model" and "build" parts of SwiftPM separate, so this PR cleans that up again.
