+++ Tests/BuildTests/BuildPlanTests.swift
@@ -26,6 +26,7 @@ import PackageLoading
 @testable import PackageModel
 
 import SPMBuildCore
+import _InternalBuildTestSupport
 import _InternalTestSupport
 import SwiftDriver
 import Workspace
